refactor: openaiService.js を services/openai/ に分割（後方互換ファサード）
501行の単一ファイルをモデル定義/設定/クライアント/プロンプト別に分割し、
openaiService.js は re-export のみの薄いファサードに変更
